pyproject.toml
[project]
name = "llm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "deep-translator>=1.11.4",
    "fastapi>=0.135.1",
    "google-genai>=1.65.0",
    "livekit>=1.1.2",
    "livekit-agents>=1.4.3",
    "livekit-api>=1.1.0",
    "livekit-plugins-groq>=1.4.3",
    "livekit-plugins-sarvam>=1.4.3",
    "livekit-plugins-silero>=1.4.3",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.41.0",
]
